Add a write() method to CPAN::Access::AdHoc::Arcive::Zip.
Also, I found that the tests for the Tar write() were not in fact being
run, and that the code was in fact buggy. Sorting this out uncovered the
fact that archive files with identical contents are not necessarily
byte-for-byte identical. So the write() tests were pulled out into
t/write.t (mostly 'cause I could skip the whole thing if the necessary
File::Temp functionality wasn't there), and had it do, for each of the
archive types, the following:

1) Read the original archive (using wrap_archive, so I didn't have to do
setup on a CPAN::Access::AdHoc class).

2) write() the archive to a temporary directory.

3) Read the rewritten archive (wrap_archive() again).

4) Do a single test comparing both archive objects, that passes only if
   the following are true:
   - base_dir() are the same;
   - both archives contain the same files;
   - corresponding files are byte-for-byte identical.


git-svn-id: file:///Users/tom/Code/svn/repos/CPAN-Access-AdHoc/trunk@3056 f6fd3633-1913-47c2-bad4-372d75c948ec