INCOMPATIBLE CHANGE * The default_cpan_source attribute is now carried internally as a   reference to an array of the expanded class names, and this is what is   returned by the accessor.
COMPATIBLE CHANGES
The mutator logic has been completely rewritten to try to facilitate
subclassing. For each attribute 'whatever', there is a
__default_whatever() method to supply the default if it is being set to
undef, and a __validate_whatever() method to validate the prospective
new value and raise an exception if there is something wrong with it.

The new() method delegates initialization of attributes to an __init()
method.

All the __* methods listed above are documented as package-private.

In addition there is an undocumented __create_accessor_mutators() static
method that creates accessor/mutator methods. This will become
documented when one of the three following happens:
* I actually write a subclass that adds an attribute;
* I come up with a better signature than this currently has;
* I get the gumption to write up its current signature.


git-svn-id: file:///Users/tom/Code/svn/repos/CPAN-Access-AdHoc/trunk@3072 f6fd3633-1913-47c2-bad4-372d75c948ec