Redo the attribute functionality in terms of three hooks for each attribute. Assuming the attribute name is 'whatever', there are: * __attr__whatever__default (provide a default value) * __attr__whatever__validate (validate or modify) * __attr__whatever__post_assignment (after value assigned)
All are optional, but the superclass provides a no-op version of the
last two if they are not found at run time.

This means that the stub mutators are no longer needed, so the code that
creates the accessor/mutator methods needs no input but the names of the
attributes. So __create_accessor_mutators() got documented, though it is
documented as private to the package.


git-svn-id: file:///Users/tom/Code/svn/repos/CPAN-Access-AdHoc/trunk@3074 f6fd3633-1913-47c2-bad4-372d75c948ec