Explicitly the permitted operator mask used by CPAN::Access::AdHoc when passing strings to Safe.
Convert the mock Safe module to use the Opcode module to mask unwanted
opcodes. It was necessary to add ':load' to get things to execute under
the coverage analyzer. Tim Bunce suggested the Opcode module in response
to my blog post "The Case of the Unsafe Safe".

In an unrelated action, flesh out the SEE ALSO section of the
CPAN::Access::AdHoc docs with a survey of modules to manipulate a CPAN
archive. There are a bunch to access specific data, but none that I can
find to handle an arbitrary file. There are ones to inject
distributions, but none that I can find to remove them. And exactly one
to manipulate the module index.


git-svn-id: file:///Users/tom/Code/svn/repos/CPAN-Access-AdHoc/trunk@3086 f6fd3633-1913-47c2-bad4-372d75c948ec