Change the CPAN::Access::AdHoc __init() method to take a reference to the arguments hash, rather than the hash itself. Also require that overrides to __init() call $self->SUPER::__init() first thing.
THIS IS AN INCOMPATIBLE CHANGE, but the old convention was poorly
thought out and did not in fact do what was intended. The superclass
must be initialized first because the subclass' mutators rely on
config() working. The arguments must be a reference, because when
the argument hash is passed unrolled, the subclass has no way to affect
the contents of the hash as seen by the superclass.


git-svn-id: file:///Users/tom/Code/svn/repos/CPAN-Access-AdHoc/trunk@3143 f6fd3633-1913-47c2-bad4-372d75c948ec