Make eg/cpan-adhoc more pipe-friendly.
If STDIN is not a terminal, we print the banner and echo the commands
only if -verbose is asserted,

If STDOUT is not a terminal, we do not issue a newline before exiting.

The corpus command was given an option, -full, which defaults to false
if there is only one argument, and true otherwise. If asserted, you get
the author name as heading, followed by indented full distribution
names. If not asserted, you get the base distribution name only.


git-svn-id: file:///Users/tom/Code/svn/repos/CPAN-Access-AdHoc/trunk@4168 f6fd3633-1913-47c2-bad4-372d75c948ec