Have Zip item_present check for item directly.
The old implementation passed a Regexp object to Archive::Zip
membersMatching(), but Regexps are badly broken under 5.19.10, at least
when used as a proxy for a string match (which this one was) with /m
asserted (ditto, though this was pro forma). It makes no sense to change
this for a broken development Perl, but using memberNamed() seems to be
more straightforward anyway.